Refacto + context manager
